+++ src/components/Radio.tsx
@@ -1,10 +1,10 @@
 import React, { HTMLProps } from 'react';
 
 interface Props {
-  id: string;
-  radioProps?: HTMLProps<HTMLDivElement>;
-  inputProps?: HTMLProps<HTMLInputElement>;
-  labelProps?: HTMLProps<HTMLLabelElement>;
+  readonly id: string;
+  readonly radioProps?: Readonly<HTMLProps<HTMLDivElement>>;
+  readonly inputProps?: Readonly<HTMLProps<HTMLInputElement>>;
+  readonly labelProps?: Readonly<HTMLProps<HTMLLabelElement>>;
 }
 
 export const Radio: React.FC<Props> = ({