+++ src/components/Textarea.tsx
@@ -1,6 +1,6 @@
 import React, { HTMLProps } from 'react';
 
-interface Props extends HTMLProps<HTMLTextAreaElement> {}
+interface Props extends Readonly<HTMLProps<HTMLTextAreaElement>> {}
 
 export const Textarea: React.FC<Props> = ({ className = '', ...props }) => (
   <textarea {...props} className={`textarea ${className}`} />