+++ CHANGELOG.md
@@ -1,3 +1,23 @@
 # React Figma UI Changelog
 
 ## 1.0.0 (2020-)
+#### 
+##### Components
+- Button
+- Checkbox
+- Disclosure
+- Icon
+- IconButton
+- Input
+- Label
+- Onboarding
+- Radio
+- SectionTitle
+- SelectMenu
+- Switch
+- Textarea
+- Type
+
+##### 
+- added Storybook preview
+- added snapshot tests with Jest